refactor(layout): finalize connections before DI emission
Finalize route docking on layout state before DI serialization, keeping the emitter geometry-neutral and avoiding per-edge peer-route reconstruction during output.
