+++ lib/Layouter.js
@@ -64,7 +64,8 @@ import {
   needsExpandedSubProcessTitleClearance
 } from './layout/LabelLayouter.js';
 import { layoutGroups } from './layout/GroupLayouter.js';
-import { emitLayout, orientPlaneDockings } from './di/LayoutEmitter.js';
+import { finalizeLayoutConnections } from './layout/ConnectionDocking.js';
+import { emitLayout } from './di/LayoutEmitter.js';
 
 /**
  * Orchestrates greenfield BPMN layout across semantic analysis, placement,
@@ -101,6 +102,7 @@ export class Layouter {
       : this.layoutScope(root);
 
     normalizeLayout(layout);
+    finalizeLayoutConnections(layout, true);
     this.emitDiagram(definitions, root, layout);
     this.emitCollapsedSubProcessDiagrams(definitions, layout);
     this.warnForMissingDi(root, definitions);
@@ -856,7 +858,6 @@ export class Layouter {
 
     definitions.diagrams.push(diagram);
     emitLayout(this.diFactory, layout, plane.planeElement);
-    orientPlaneDockings(this.diFactory, plane.planeElement);
     layoutExternalLabels(this.diFactory, plane.planeElement);
   }
 
@@ -868,6 +869,7 @@ export class Layouter {
       }
 
       normalizeLayout(child);
+      finalizeLayoutConnections(child);
       const plane = this.diFactory.createDiPlane({
         id: `BPMNPlane_${child.scope.id}`,
         bpmnElement: child.scope,
